Extract biquad logic to separate file
